Handle the deleting on one question using sweet alert

+++ src/components/QAAccordion.jsx
@@ -1,13 +1,32 @@
 import Accordion from 'react-bootstrap/Accordion';
+import Swal from 'sweetalert2';
 
 function QAAccordion({ data, setQuestions, notify }) {
     const deleteQuestionHandler = () => {
-        setQuestions((prev) => {
-            const prevData = prev.filter((item) => item.id !== data.id);
-            localStorage.setItem('QUESTIONS', JSON.stringify(prevData));
-            return prevData;
+        Swal.fire({
+            title: 'هل انت متاكد من حذف السؤال؟',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#224f61',
+            cancelButtonColor: '#dc3545',
+            confirmButtonText: 'نعم',
+            cancelButtonText: 'لا',
+            customClass: {
+                title: 'font',
+                content: 'font',
+                confirmButton: 'font',
+                cancelButton: 'font',
+            },
+        }).then((result) => {
+            if (result.isConfirmed) {
+                setQuestions((prev) => {
+                    const prevData = prev.filter((item) => item.id !== data.id);
+                    localStorage.setItem('QUESTIONS', JSON.stringify(prevData));
+                    return prevData;
+                });
+                notify('تم حذف السؤال بنجاح', 'success');
+            }
         });
-        notify('تم حذف السؤال بنجاح', 'success');
     };
     return (
         <Accordion className="my-3">
